// src/app/e-wallet/models/api/virtualCardCreateModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface VirtualCardCreateModel { 
    readonly accountId?: string;
    readonly productId?: string;
    readonly cardName: string;
}